refactor: Relocate resource link launch code
fix:  Launch message verification order
refactor: Rename and relocate get_resource_launch method
fix:  Code quality and tests
