+++ openedx_lti_tool_plugin/resource_link_launch/__init__.py
@@ -1,0 +1,8 @@
+"""LTI Resource Link Launch.
+
+This module handles all the code related to resource link launch requests.
+
+.. _LTI Core Specification 1.3 - Resource link launch request message:
+    https://www.imsglobal.org/spec/lti/v1p3#resource-link-launch-request-message
+
+"""
