+++ openedx_lti_tool_plugin/resource_link_launch/ags/__init__.py
@@ -1,0 +1,13 @@
+"""LTI Assignment and Grade Services.
+
+This module handles all the code related to LTI AGS (Assignment and Grade Services).
+
+Attributes:
+    MODULE_PATH (str): This module absolute path.
+
+.. _LTI Assignment and Grade Services Specification:
+    https://www.imsglobal.org/spec/lti-ags/v2p0
+
+"""
+
+MODULE_PATH = 'openedx_lti_tool_plugin.resource_link_launch.ags'
