+++ openedx_lti_tool_plugin/resource_link_launch/ags/tests/__init__.py
@@ -1,0 +1,3 @@
+"""Test ags module."""
+
+MODULE_PATH = 'openedx_lti_tool_plugin.resource_link_launch.ags'
