+++ openedx_lti_tool_plugin/resource_link_launch/ags/tests/test_tasks.py
@@ -1,39 +1,42 @@
-"""Tests for the openedx_lti_tool_plugin tasks module."""
+"""Tests tasks module."""
 from unittest.mock import MagicMock, patch
 
 from django.test import TestCase
 from testfixtures import log_capture
 from testfixtures.logcapture import LogCaptureForDecorator
 
-from openedx_lti_tool_plugin.tasks import send_problem_score_update, send_vertical_score_update
+from openedx_lti_tool_plugin.resource_link_launch.ags.tasks import send_problem_score_update, send_vertical_score_update
+from openedx_lti_tool_plugin.resource_link_launch.ags.tests import MODULE_PATH
 from openedx_lti_tool_plugin.tests import COURSE_ID, USAGE_KEY
+
+MODULE_PATH = f'{MODULE_PATH}.tasks'
 
 
 class TestSendVerticalScoreUpdate(TestCase):
-    """Test send_vertical_score_update task."""
+    """Test send_vertical_score_update function."""
 
     def setUp(self):
-        """Test fixtures setup."""
+        """Set up test fixtures."""
+        self.user = MagicMock()
         self.user_id = 1
         self.course_id = COURSE_ID
-        self.problem_id = USAGE_KEY
-        self.user = MagicMock()
-        self.problem_descriptor = MagicMock()
         self.course_descriptor = MagicMock()
         self.course_grade = MagicMock()
         self.course_grade.score_for_module.return_value = (1, 1)
+        self.problem_id = USAGE_KEY
+        self.problem_descriptor = MagicMock()
         self.vertical_graded_resource = MagicMock()
 
     @log_capture()
-    @patch('openedx_lti_tool_plugin.tasks.get_user_model')
-    @patch('openedx_lti_tool_plugin.tasks.UsageKey')
-    @patch('openedx_lti_tool_plugin.tasks.modulestore')
-    @patch('openedx_lti_tool_plugin.tasks.LtiGradedResource')
-    @patch('openedx_lti_tool_plugin.tasks.CourseKey')
-    @patch('openedx_lti_tool_plugin.tasks.course_grade_factory')
-    @patch('openedx_lti_tool_plugin.tasks.datetime')
-    @patch('openedx_lti_tool_plugin.tasks.timezone')
-    def test_send_vertical_score_update(
+    @patch(f'{MODULE_PATH}.get_user_model')
+    @patch(f'{MODULE_PATH}.UsageKey')
+    @patch(f'{MODULE_PATH}.modulestore')
+    @patch(f'{MODULE_PATH}.LtiGradedResource')
+    @patch(f'{MODULE_PATH}.CourseKey')
+    @patch(f'{MODULE_PATH}.course_grade_factory')
+    @patch(f'{MODULE_PATH}.datetime')
+    @patch(f'{MODULE_PATH}.timezone')
+    def test_with_vertical_score_update(
         self,
         timezone_mock: MagicMock,
         datetime_mock: MagicMock,
@@ -45,19 +48,7 @@ class TestSendVerticalScoreUpdate(TestCase):
         get_user_model_mock: MagicMock,
         log: LogCaptureForDecorator,
     ):
-        """Test send_vertical_score_update task.
-
-        Args:
-            timezone_mock: mocked timezone module.
-            datetime_mock: mocked datetime module.
-            course_grade_factory_mock: mocked CourseGradeFactory class.
-            course_key_mock: mocked CourseKey class.
-            lti_graded_resource_mock: mocked LtiGradedResource model.
-            modulestore_mock: mocked modulestore.
-            usage_key_mock: mocked UsageKey class.
-            get_user_model_mock: mocked get_user_model function.
-            log: LogCapture fixture.
-        """
+        """Test with vertical score update."""
         get_user_model_mock.return_value.objects.get.return_value = self.user
         vertical_key = MagicMock()
         self.problem_descriptor.parent = vertical_key
@@ -79,14 +70,16 @@ class TestSendVerticalScoreUpdate(TestCase):
         get_user_model_mock.assert_called_once_with()
         get_user_model_mock.return_value.objects.get.assert_called_once_with(id=self.user_id)
         usage_key_mock.from_string.assert_called_once_with(self.problem_id)
-        modulestore_mock.return_value.get_item.assert_called_once_with(usage_key_mock.from_string.return_value)
+        modulestore_mock.return_value.get_item.assert_called_once_with(
+            usage_key_mock.from_string.return_value,
+        )
         lti_graded_resource_mock.objects.all_from_user_id.assert_called_once_with(
             user_id=self.user.id,
             context_key=str(self.problem_descriptor.parent),
         )
         log.check(
             (
-                'openedx_lti_tool_plugin.tasks',
+                MODULE_PATH,
                 'INFO',
                 f'LTI AGS: Sending AGS update for unit {vertical_key} with user {self.user_id}',
             ),
@@ -106,15 +99,15 @@ class TestSendVerticalScoreUpdate(TestCase):
             datetime_mock.now.return_value,
         )
 
-    @patch('openedx_lti_tool_plugin.tasks.log')
-    @patch('openedx_lti_tool_plugin.tasks.get_user_model')
-    @patch('openedx_lti_tool_plugin.tasks.UsageKey')
-    @patch('openedx_lti_tool_plugin.tasks.modulestore')
-    @patch('openedx_lti_tool_plugin.tasks.LtiGradedResource')
-    @patch('openedx_lti_tool_plugin.tasks.CourseKey')
-    @patch('openedx_lti_tool_plugin.tasks.course_grade_factory')
-    @patch('openedx_lti_tool_plugin.tasks.datetime')
-    def test_send_vertical_score_update_without_vertical_resources(
+    @patch(f'{MODULE_PATH}.log')
+    @patch(f'{MODULE_PATH}.get_user_model')
+    @patch(f'{MODULE_PATH}.UsageKey')
+    @patch(f'{MODULE_PATH}.modulestore')
+    @patch(f'{MODULE_PATH}.LtiGradedResource')
+    @patch(f'{MODULE_PATH}.CourseKey')
+    @patch(f'{MODULE_PATH}.course_grade_factory')
+    @patch(f'{MODULE_PATH}.datetime')
+    def test_without_graded_resources(
         self,
         datetime_mock: MagicMock,
         course_grade_factory_mock: MagicMock,
@@ -125,18 +118,7 @@ class TestSendVerticalScoreUpdate(TestCase):
         get_user_model_mock: MagicMock,
         log_mock: MagicMock,
     ):
-        """Test send_vertical_score_update task without graded resources for vertical.
-
-        Args:
-            datetime_mock: mocked datetime module.
-            course_grade_factory_mock: mocked CourseGradeFactory class.
-            course_key_mock: mocked CourseKey class.
-            lti_graded_resource_mock: mocked LtiGradedResource model.
-            modulestore_mock: mocked modulestore.
-            usage_key_mock: mocked UsageKey class.
-            get_user_model_mock: mocked get_user_model function.
-            log_mock: mocked log.
-        """
+        """Test without graded resources."""
         get_user_model_mock.return_value.objects.get.return_value = self.user
         modulestore_mock.return_value.get_item.return_value = self.problem_descriptor
         course_grade_factory_mock.return_value.read.return_value = self.course_grade
@@ -161,36 +143,29 @@ class TestSendVerticalScoreUpdate(TestCase):
         self.vertical_graded_resource.update_score.assert_not_called()
 
 
-@patch('openedx_lti_tool_plugin.tasks.LtiGradedResource')
+@patch(f'{MODULE_PATH}.LtiGradedResource')
 class TestSendProblemScoreUpdate(TestCase):
-    """Test send_problem_score_update task."""
+    """Test send_problem_score_update function."""
 
     def setUp(self):
-        """Test fixtures setup."""
-        self.problem_weighted_earned = 1
-        self.problem_weighted_possible = 1
+        """Set up test fixtures."""
         self.user_id = 1
         self.problem_id = USAGE_KEY
+        self.problem_weighted_earned = 1
+        self.problem_weighted_possible = 1
         self.graded_resource = MagicMock()
 
     @log_capture()
-    @patch('openedx_lti_tool_plugin.tasks.datetime')
-    @patch('openedx_lti_tool_plugin.tasks.timezone')
-    def test_send_problem_score_update(
+    @patch(f'{MODULE_PATH}.datetime')
+    @patch(f'{MODULE_PATH}.timezone')
+    def test_with_problem_score_update(
         self,
         timezone_mock: MagicMock,
         datetime_mock: MagicMock,
         log: LogCaptureForDecorator,
         lti_graded_resource_mock: MagicMock,
     ):
-        """Test send_problem_score_update task.
-
-        Args:
-            timezone_mock: mocked timezone module.
-            datetime_mock: mocked datetime module.
-            log: LogCapture fixture.
-            lti_graded_resource_mock: mocked LtiGradedResource model.
-        """
+        """Test with problem score update."""
         lti_graded_resource_mock.objects.all_from_user_id.return_value = [
             self.graded_resource,
         ]
@@ -210,7 +185,7 @@ class TestSendProblemScoreUpdate(TestCase):
         )
         log.check(
             (
-                'openedx_lti_tool_plugin.tasks',
+                MODULE_PATH,
                 'INFO',
                 f'LTI AGS: Sending AGS update for problem {self.problem_id} with user {self.user_id}',
             ),
@@ -222,18 +197,13 @@ class TestSendProblemScoreUpdate(TestCase):
             datetime_mock.now.return_value,
         )
 
-    @patch('openedx_lti_tool_plugin.tasks.log')
-    def test_send_problem_score_update_without_graded_resource(
+    @patch(f'{MODULE_PATH}.log')
+    def test_without_graded_resource(
         self,
         log_mock: MagicMock,
         lti_graded_resource_mock: MagicMock,
     ):
-        """Test send_problem_score_update task without graded resources.
-
-        Args:
-            log_mock: mocked log.
-            lti_graded_resource_mock: mocked LtiGradedResource model.
-        """
+        """Test without graded resources."""
         lti_graded_resource_mock.objects.all_from_user_id.return_value = []
 
         self.assertEqual(
