+++ openedx_lti_tool_plugin/resource_link_launch/exceptions.py
@@ -1,4 +1,4 @@
-"""Exceptions for openedx_lti_tool_plugin."""
+"""Exceptions."""
 
 
 class LtiToolLaunchException(Exception):
