+++ openedx_lti_tool_plugin/resource_link_launch/tests/__init__.py
@@ -1,0 +1,3 @@
+"""Test resource_link_launch module."""
+
+MODULE_PATH = 'openedx_lti_tool_plugin.resource_link_launch'
