+++ openedx_lti_tool_plugin/tests/test_admin.py
@@ -1,20 +1,12 @@
 """Tests for the openedx_lti_tool_plugin admin module."""
-from django.contrib import admin
 from django.contrib.admin.sites import AdminSite
 from django.contrib.auth import get_user_model
 from django.test import TestCase
-from django.urls import path, reverse
 from pylti1p3.contrib.django.lti1p3_tool_config.models import LtiTool, LtiToolKey
 
-from openedx_lti_tool_plugin.admin import CourseAccessConfigurationAdmin
+from openedx_lti_tool_plugin.admin import CourseAccessConfigurationAdmin, LtiProfileAdmin
 from openedx_lti_tool_plugin.models import CourseAccessConfiguration, LtiProfile
 from openedx_lti_tool_plugin.tests import AUD, ISS, SUB
-from openedx_lti_tool_plugin.urls import urlpatterns
-
-
-def get_admin_view_url(obj: LtiProfile, name: str) -> str:
-    """Get admin URL from model instance."""
-    return f'admin:{obj._meta.app_label}_{type(obj).__name__.lower()}_{name}'
 
 
 class TestLtiProfileAdmin(TestCase):
@@ -22,28 +14,15 @@ class TestLtiProfileAdmin(TestCase):
 
     def setUp(self):
         """Test fixtures setup."""
+        self.admin = LtiProfileAdmin(LtiProfile, AdminSite)
         self.user = get_user_model().objects.create_superuser(username='x', password='x', email='x@example.com')
         self.client.force_login(self.user)
         self.profile = LtiProfile.objects.create(platform_id=ISS, client_id=AUD, subject_id=SUB)
-        urlpatterns.append(path('admin/', admin.site.urls))
 
-    def test_add_view(self):
-        """Test admin add view can be reached."""
-        url = reverse(get_admin_view_url(self.profile, 'add'))
-
-        self.assertEqual(self.client.get(url).status_code, 200)
-
-    def test_change_view(self):
-        """Test admin change view can be reached."""
-        url = reverse(get_admin_view_url(self.profile, 'change'), args=(self.profile.pk,))
-
-        self.assertEqual(self.client.get(url).status_code, 200)
-
-    def test_delete_view(self):
-        """Test admin delete view can be reached."""
-        url = reverse(get_admin_view_url(self.profile, 'delete'), args=(self.profile.pk,))
-
-        self.assertEqual(self.client.get(url).status_code, 200)
+    def test_instance_attributes(self):
+        """Test instance attributes."""
+        self.assertEqual(self.admin.list_display, ('id', 'uuid', 'platform_id', 'client_id', 'subject_id'))
+        self.assertEqual(self.admin.search_fields, ['id', 'uuid', 'platform_id', 'client_id', 'subject_id'])
 
 
 class TestCourseAccessConfigurationAdmin(TestCase):
@@ -51,6 +30,7 @@ class TestCourseAccessConfigurationAdmin(TestCase):
 
     def setUp(self):
         """Test fixtures setup."""
+        self.admin = CourseAccessConfigurationAdmin(CourseAccessConfiguration, AdminSite)
         self.user = get_user_model().objects.create_superuser(username='x', password='x', email='x@example.com')
         self.client.force_login(self.user)
         self.lti_tool = LtiTool.objects.create(
@@ -62,28 +42,12 @@ class TestCourseAccessConfigurationAdmin(TestCase):
             tool_key=LtiToolKey.objects.create(),
         )
         self.access_configuration = CourseAccessConfiguration.objects.get(lti_tool=self.lti_tool)
-        urlpatterns.append(path('admin/', admin.site.urls))
 
-    def test_add_view(self):
-        """Test admin add view can be reached."""
-        url = reverse(get_admin_view_url(self.access_configuration, 'add'))
-
-        self.assertEqual(self.client.get(url).status_code, 200)
-
-    def test_change_view(self):
-        """Test admin change view can be reached."""
-        url = reverse(get_admin_view_url(self.access_configuration, 'change'), args=(self.access_configuration.pk,))
-
-        self.assertEqual(self.client.get(url).status_code, 200)
-
-    def test_delete_view(self):
-        """Test admin delete view can be reached."""
-        url = reverse(get_admin_view_url(self.access_configuration, 'delete'), args=(self.access_configuration.pk,))
-
-        self.assertEqual(self.client.get(url).status_code, 200)
+    def test_instance_attributes(self):
+        """Test instance attributes."""
+        self.assertEqual(self.admin.list_display, ('id', 'lti_tool_title', 'allowed_course_ids'))
+        self.assertEqual(self.admin.search_fields, ['id', 'lti_tool__title', 'allowed_course_ids'])
 
     def test_lti_tool_title(self):
         """Test lti_tool_title method."""
-        admin_cls = CourseAccessConfigurationAdmin(CourseAccessConfiguration, AdminSite)
-
-        self.assertEqual(admin_cls.lti_tool_title(self.access_configuration), 'random-title')
+        self.assertEqual(self.admin.lti_tool_title(self.access_configuration), 'random-title')
