+++ openedx_lti_tool_plugin/tests/test_models.py
@@ -10,17 +10,12 @@ from opaque_keys.edx.keys import CourseKey
 from pylti1p3.contrib.django.lti1p3_tool_config.models import LtiTool, LtiToolKey
 
 from openedx_lti_tool_plugin.apps import OpenEdxLtiToolPluginConfig as app_config
-from openedx_lti_tool_plugin.models import (
-    CourseAccessConfiguration,
-    LtiGradedResource,
-    LtiGradedResourceManager,
-    LtiProfile,
-)
+from openedx_lti_tool_plugin.models import CourseAccessConfiguration, LtiProfile
 from openedx_lti_tool_plugin.tests import AUD, ISS, SUB
 
 
-class LtiProfileMixin():
-    """Add LTI profile instance to test fixtures setup."""
+class TestLtiProfile(TestCase):
+    """Test LTI 1.3 profile model."""
 
     def setUp(self):
         """Add RequestFactory to test setup."""
@@ -32,10 +27,6 @@ class LtiProfileMixin():
             subject_id=SUB,
             pii=self.pii
         )
-
-
-class TestLtiProfile(LtiProfileMixin, TestCase):
-    """Test LTI 1.3 profile model."""
 
     @patch.object(get_user_model().objects, 'get_or_create')
     @patch('openedx_lti_tool_plugin.models.getattr', return_value=False)
@@ -347,107 +338,3 @@ class TestCourseAccessConfiguration(TestCase):
             str(self.access_configuration),
             f'<CourseAccessConfiguration, ID: {self.access_configuration.id}>',
         )
-
-
-class TestLtiGradedResourceManager(TestCase):
-    """Test LTI 1.3 profile model manager."""
-
-    @patch.object(LtiGradedResourceManager, 'filter')
-    @patch.object(LtiProfile.objects, 'filter')
-    def test_all_from_user_id(
-        self,
-        lti_profile_filter_mock: MagicMock,
-        graded_resource_filter_mock: MagicMock,
-    ):
-        """Test LtiGradedResourceManager all_from_user_id method.
-
-        Args:
-            lti_profile_filter_mock: Mocked LtiProfile filter method.
-            graded_resource_filter_mock: Mocked LtiGradedResourceManager filter method.
-        """
-        result = LtiGradedResource.objects.all_from_user_id(user_id='random-user-id', context_key='random-key')
-
-        lti_profile_filter_mock.assert_called_once_with(user__id='random-user-id')
-        lti_profile_filter_mock.return_value.first.assert_called_once_with()
-        graded_resource_filter_mock.assert_called_once_with(
-            lti_profile=lti_profile_filter_mock().first(),
-            context_key='random-key',
-        )
-        self.assertEqual(result, graded_resource_filter_mock())
-
-
-class TestLtiGradedResource(LtiProfileMixin, TestCase):
-    """Test LTI graded resource model."""
-
-    def setUp(self):
-        """Test fixtures setup."""
-        super().setUp()
-        self.lti_graded_resource = LtiGradedResource.objects.create(
-            lti_profile=self.profile,
-            context_key='course-v1:test+test+test',
-            lineitem='https://random-lineitem.test',
-        )
-
-    @patch('openedx_lti_tool_plugin.models.Grade')
-    @patch('openedx_lti_tool_plugin.models.DjangoMessageLaunch')
-    @patch('openedx_lti_tool_plugin.models.DjangoDbToolConf')
-    def test_update_score(
-        self,
-        tool_conf_mock: MagicMock,
-        message_launch_mock: MagicMock,
-        grade_mock: MagicMock,
-    ):
-        """Test update_score method.
-
-        Args:
-            tool_conf_mock: Mocked DjangoDbToolConf class.
-            message_launch_mock: Mocked DjangoMessageLaunch class.
-            grade_mock: Mocked Grade class.
-        """
-        timestamp = MagicMock()
-        timestamp.isoformat.return_value = 'random-timestamp'
-
-        self.lti_graded_resource.update_score('random-given-score', 'random-max-score', timestamp)
-
-        tool_conf_mock.assert_called_once_with()
-        message_launch_mock.assert_called_once_with(request=None, tool_config=tool_conf_mock())
-        message_launch_mock().set_auto_validation.assert_called_once_with(enable=False)
-        message_launch_mock().set_jwt.assert_called_once_with({
-            'body': {
-                'iss': self.profile.platform_id,
-                'aud': self.profile.client_id,
-                'https://purl.imsglobal.org/spec/lti-ags/claim/endpoint': {
-                    'lineitem': 'https://random-lineitem.test',
-                    'scope': {
-                        'https://purl.imsglobal.org/spec/lti-ags/scope/lineitem',
-                        'https://purl.imsglobal.org/spec/lti-ags/scope/score',
-                    },
-                },
-            },
-        })
-        message_launch_mock().set_restored.assert_called_once_with()
-        message_launch_mock().validate_registration.assert_called_once_with()
-        message_launch_mock().get_ags.assert_called_once_with()
-        grade_mock.assert_called_once_with()
-        grade_mock().set_score_given.assert_called_once_with('random-given-score')
-        grade_mock().set_score_maximum.assert_called_once_with('random-max-score')
-        grade_mock().set_timestamp.assert_called_once_with('random-timestamp')
-        grade_mock().set_activity_progress.assert_called_once_with('Submitted')
-        grade_mock().set_grading_progress.assert_called_once_with('FullyGraded')
-        grade_mock().set_user_id.assert_called_once_with(self.profile.subject_id)
-        message_launch_mock().get_ags().put_grade.asser_called_once_with(grade_mock().set_user_id())
-
-    def test_str_method(self):
-        """Test __str__ method return value."""
-        self.assertEqual(str(self.lti_graded_resource), f'<LtiGradedResource, ID: {self.lti_graded_resource.id}>')
-
-    @patch.object(LtiGradedResource, 'full_clean')
-    def test_save(self, full_clean_mock: MagicMock):
-        """Test save method."""
-        LtiGradedResource(
-            lti_profile=self.profile,
-            context_key='course-v1:test+test+test',
-            lineitem='https://random-lineitem2.test',
-        ).save()
-
-        full_clean_mock.assert_called_once_with()
