--- openedx_lti_tool_plugin/tests/views/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the `openedx_lti_tool_plugin.views.LtiToolLaunchView` class."""
