+++ openedx_lti_tool_plugin/urls.py
@@ -1,22 +1,13 @@
 """URL configuration for openedx_lti_tool_plugin."""
-from django.conf import settings
-from django.urls import include, path, re_path
+from django.urls import include, path
 
 from openedx_lti_tool_plugin import views
 
+lti_1p3_urls = [
+    path('login', views.LtiToolLoginView.as_view(), name='login'),
+    path('pub/jwks', views.LtiToolJwksView.as_view(), name='jwks'),
+    path('launch/', include('openedx_lti_tool_plugin.resource_link_launch.urls')),
+]
 urlpatterns = [
-    path('1.3/', include([
-        path('login', views.LtiToolLoginView.as_view(), name='lti1p3-login'),
-        path('pub/jwks', views.LtiToolJwksView.as_view(), name='lti1p3-pub-jwks'),
-        re_path(
-            fr'^launch/{settings.COURSE_ID_PATTERN}$',
-            views.LtiToolLaunchView.as_view(),
-            name='lti1p3-launch',
-        ),
-        re_path(
-            fr'^launch/{settings.COURSE_ID_PATTERN}/{settings.USAGE_KEY_PATTERN}$',
-            views.LtiToolLaunchView.as_view(),
-            name='lti1p3-launch',
-        ),
-    ])),
+    path('1.3/', include((lti_1p3_urls, '1.3'))),
 ]
